Update token regardless of the action
This is due to how the session system works. When an action is performed, it needs to verify that the provided token is valid, and, by doing so, it is refreshed if needed. That's why the client has to, constantly, update it to a new value to prevent desynchronisations with the server.