Force redraw when a component is updated
This is done to fix some problems when components did not show the updated content properly, like the object tag.
Achieved by setting the component's content to null before updating it. Then, drawing a display: none div if the previous case is given.
That forces the browser to redraw it with the new content, once it's updated in the state.