Made advancements into the draggable mode for `Dashboard`
Now the draggable elements from the components are only draggable per se
once the "reordering" mode is established. This prevents future
accidental reorders when clicking the components to edit, and as well
enabling future mobile reorders in place.

Also, the UI shown to toggle the reorder mode is an obvious WIP, since
it's not even styled yet /shrug